# Minimal working example with PSDMatrices.jl