--- a/src/ProbNumDiffEq.jl
+++ b/src/ProbNumDiffEq.jl
@@ -64,6 +64,8 @@ _matmul!(
 const SRMatrix = PSDMatrix
 export SRMatrix
 diag(S::SRMatrix) = map(c -> sum(abs2, c), eachcol(S.R))
+X_A_Xt(A, X) = X * A * X'
+X_A_Xt!(out, A, X) = (out .= X * A * X')
 apply_diffusion(Q, diffusion::Diagonal) = X_A_Xt(Q, sqrt.(diffusion))
 apply_diffusion(Q::SRMatrix, diffusion::Number) = SRMatrix(sqrt.(diffusion) * Q.R)