--- a/src/ProbNumDiffEq.jl
+++ b/src/ProbNumDiffEq.jl
@@ -114,19 +114,19 @@ export ManifoldUpdate
 
 # Do as they do here:
 # https://github.com/SciML/OrdinaryDiffEq.jl/blob/v5.61.1/src/OrdinaryDiffEq.jl#L175-L193
-# let
-#     while true
-#         function lorenz(du, u, p, t)
-#             du[1] = 10.0(u[2] - u[1])
-#             du[2] = u[1] * (28.0 - u[3]) - u[2]
-#             return du[3] = u[1] * u[2] - (8 / 3) * u[3]
-#         end
+let
+    while true
+        function lorenz(du, u, p, t)
+            du[1] = 10.0(u[2] - u[1])
+            du[2] = u[1] * (28.0 - u[3]) - u[2]
+            return du[3] = u[1] * u[2] - (8 / 3) * u[3]
+        end
 
-#         lorenzprob = ODEProblem(lorenz, [1.0; 0.0; 0.0], (0.0, 1.0))
-#         solve(lorenzprob, EK0())
-#         solve(lorenzprob, EK1())
-#         break
-#     end
-# end
+        lorenzprob = ODEProblem(lorenz, [1.0; 0.0; 0.0], (0.0, 1.0))
+        solve(lorenzprob, EK0())
+        solve(lorenzprob, EK1())
+        break
+    end
+end
 
 end