--- a/src/caches.jl
+++ b/src/caches.jl
@@ -126,10 +126,7 @@ function OrdinaryDiffEq.alg_cache(
     A, Q = ibm(d, q, uElType)
 
     initial_variance = ones(uElType, D)
-    x0 = Gaussian(
-        zeros(uElType, D),
-        SRMatrix(diagm(sqrt.(initial_variance))),
-    )
+    x0 = Gaussian(zeros(uElType, D), SRMatrix(diagm(sqrt.(initial_variance))))
 
     # Measurement model
     R = zeros(uElType, d, d)