--- a/src/callbacks.jl
+++ b/src/callbacks.jl
@@ -7,7 +7,7 @@ function manifoldupdate!(integ, residualf; maxiters=100, ϵ₁=1e-25, ϵ₂=1e-1
     result = DiffResults.JacobianResult(z_tmp, tmp)
     d = length(z_tmp)
     H = H[1:d, :]
-    S = SquarerootMatrix(C.squareroot[1:d, :], C.mat[1:d, 1:d])
+    S = SRMatrix(C.R[:, 1:d])
     m_tmp, C_tmp = x_tmp
 
     m_i = copy(m)
@@ -24,7 +24,7 @@ function manifoldupdate!(integ, residualf; maxiters=100, ϵ₁=1e-25, ϵ₂=1e-1
 
         # m_i_new, C_i_new = update(x, Gaussian(z .+ (H * (m - m_i)), S), H)
         # More efficient update with less allocations:
-        K = C * (H' / S)
+        K = C.R' * C.R * (H' / S)
         m_tmp .= m_i .- m
         mul!(z_tmp, H, m_tmp)
         z_tmp .-= z