--- a/src/filtering/predict.jl
+++ b/src/filtering/predict.jl
@@ -13,7 +13,7 @@ predict(x::Gaussian, A::AbstractMatrix, Q::AbstractMatrix) =
 predict_mean(x::Gaussian, A::AbstractMatrix) = A * x.μ
 predict_cov(x::Gaussian, A::AbstractMatrix, Q::AbstractMatrix) = A * x.Σ * A' + Q
 predict_cov(x::SRGaussian, A::AbstractMatrix, Q::SRMatrix) =
-    SRMatrix(qr([A * x.Σ.squareroot Q.squareroot]').R')
+    SRMatrix(qr([x.Σ.R * A'; Q.R]).R)
 
 """
     predict!(x_out, x_curr, Ah, Qh, cachemat)
@@ -21,7 +21,7 @@ predict_cov(x::SRGaussian, A::AbstractMatrix, Q::SRMatrix) =
 In-place and square-root implementation of [`predict`](@ref)
 which saves the result into `x_out`.
 
-Only works with `ProbNumDiffEq.SquarerootMatrix` types as `Ah`, `Qh`, and in the
+Only works with `PSDMatrix` types as `Ah`, `Qh`, and in the
 covariances of `x_curr` and `x_out` (both of type `Gaussian`).
 To prevent allocations, a cache matrix `cachemat` of size ``D \\times 2D``
 (where ``D \\times D`` is the size of `Ah` and `Qh`) needs to be passed.
@@ -54,19 +54,17 @@ function predict_cov!(
     cachemat::SRMatrix,
     diffusion=1,
 )
-    M, L = cachemat.mat, cachemat.squareroot
-    D, D = size(Qh.mat)
+    L = cachemat.R'
+    M = Matrix(cachemat)
+    D, D = size(Qh)
 
-    _matmul!(view(L, 1:D, 1:D), Ah, x_curr.Σ.squareroot)
-    _matmul!(view(L, 1:D, D+1:2D), sqrt.(diffusion), Qh.squareroot)
+    _matmul!(view(L, 1:D, 1:D), Ah, x_curr.Σ.R')
+    _matmul!(view(L, 1:D, D+1:2D), sqrt.(diffusion), Qh.R')
     _matmul!(M, L, L')
     chol = cholesky!(Symmetric(M), check=false)
 
-    QL =
-        issuccess(chol) ? Matrix(chol.U)' :
-        eltype(L) <: Union{Float16,Float32,Float64} ? lq!(L).L : qr(L').R'
-    # QL = eltype(L) <: Union{Float16,Float32,Float64} ? lq!(L).L : qr(L').R'
-    copy!(x_out.Σ.squareroot, QL)
-    _matmul!(x_out.Σ.mat, QL, QL')
+    Q_R = issuccess(chol) ? Matrix(chol.U) : qr(L').R
+    copy!(x_out.Σ.R, Q_R)
+    # _matmul!(x_out.Σ.mat, QL, QL')
     return x_out.Σ
 end