… Project.toml
… src/filtering/smooth.jl
… src/filtering/update.jl
… src/gaussians.jl
… src/initialization/classicsolverinit.jl
… src/initialization/common.jl
… src/initialization/taylormode.jl
… src/perform_step.jl
… src/priors.jl
… src/solution.jl
… src/solution_plotting.jl
… src/solution_sampling.jl
… src/squarerootmatrix.jl
… test/filtering.jl
… test/preconditioning.jl
… test/priors.jl
… test/solution.jl
… test/specific_problems.jl
… test/state_init.jl