This PR should solve #120.

Right now the code runs, but it allocates much more than before (since I often just replaces `M.mat` with `Matrix(M)` to get it to run). This should also be fixed in this PR.