@nathanaelbosch 2022-03-22
To get a feeling for the current performance change: 

before:
```
julia> @btime solve(prob, EK0());
  12.271 ms (45245 allocations: 4.78 MiB)

julia> @btime solve(prob, EK1());
  11.833 ms (29064 allocations: 4.33 MiB)

julia> @btime solve(prob, EK0(smooth=false), dense=false);
  5.196 ms (16424 allocations: 1.15 MiB)

julia> @btime solve(prob, EK1(smooth=false), dense=false);
  5.248 ms (17074 allocations: 1.18 MiB)
```

after
```
julia> @btime solve(prob, EK0());
  26.002 ms (45552 allocations: 79.92 MiB)

julia> @btime solve(prob, EK1());
  26.336 ms (46913 allocations: 79.74 MiB)

julia> @btime solve(prob, EK0(smooth=false), dense=false);
  8.546 ms (21812 allocations: 17.29 MiB)

julia> @btime solve(prob, EK1(smooth=false), dense=false);
  8.997 ms (23240 allocations: 17.29 MiB)
```

Calls to `plot(sol)` also became much slower (but the plotting is quite badly implemented in general tbh).