@nathanaelbosch 2022-05-09
This PR is finally faster than `main`! :tada: 

Old:
```
julia> @btime solve(prob, EK0());
  3.878 ms (15084 allocations: 1.64 MiB)

julia> @btime solve(prob, EK1());
  3.260 ms (9192 allocations: 1.36 MiB)

julia> @btime solve(prob, EK0(smooth=false), dense=false);
  1.671 ms (6324 allocations: 718.53 KiB)

julia> @btime solve(prob, EK1(smooth=false), dense=false);
  1.558 ms (6163 allocations: 667.88 KiB)
```

New:
```
julia> @btime solve(prob, EK0());
  2.464 ms (8062 allocations: 1.47 MiB)

julia> @btime solve(prob, EK1());
  2.241 ms (7518 allocations: 1.33 MiB)

julia> @btime solve(prob, EK0(smooth=false), dense=false);
  1.399 ms (5073 allocations: 733.97 KiB)

julia> @btime solve(prob, EK1(smooth=false), dense=false);
  1.279 ms (4845 allocations: 673.55 KiB)
```